auto installation of nodemon while setup via npx
